requirements.txt
streamlit
supabase
python-docx
math